Reading policy from zookeeper when running in mesos
