+++ driver/src/main/scala/com/stratio/sparkta/driver/SparktaJob.scala
@@ -19,32 +19,30 @@ package com.stratio.sparkta.driver
 
 import java.io._
 import java.nio.file.{Files, Paths}
-
-import org.json4s.DefaultFormats
-import org.json4s.ext.EnumNameSerializer
-
-import scala.Predef
 import scala.annotation.tailrec
 import scala.collection.JavaConversions._
 import scala.util._
 
 import akka.event.slf4j.SLF4JLogging
+import org.apache.commons.io.FileUtils
+import org.apache.spark.sql.SQLContext
+import org.apache.spark.streaming.dstream.DStream
+import org.apache.spark.streaming.{Duration, StreamingContext}
+import org.apache.spark.{SparkConf, SparkContext}
+import org.json4s.DefaultFormats
+import org.json4s.ext.EnumNameSerializer
+import org.json4s.native.Serialization._
+import org.reflections.Reflections
+
 import com.stratio.sparkta.aggregator.{Cube, CubeMaker}
 import com.stratio.sparkta.driver.exception.DriverException
 import com.stratio.sparkta.driver.factory.{SchemaFactory, SparkContextFactory}
-import com.stratio.sparkta.driver.models.{StreamingContextStatusEnum, AggregationPoliciesModel, OperatorModel}
 import com.stratio.sparkta.driver.service.RawDataStorageService
-import com.stratio.sparkta.sdk._
 import com.stratio.sparkta.sdk.TypeOp.TypeOp
 import com.stratio.sparkta.sdk.WriteOp.WriteOp
+import com.stratio.sparkta.sdk._
+import com.stratio.sparkta.serving.core.models._
 import com.stratio.sparkta.serving.core.{AppConstant, CuratorFactoryHolder, SparktaConfig}
-import org.apache.commons.io.FileUtils
-import org.apache.spark.{SparkConf, SparkContext}
-import org.apache.spark.sql.SQLContext
-import org.apache.spark.streaming.{Duration, StreamingContext}
-import org.apache.spark.streaming.dstream.DStream
-import org.json4s.native.Serialization._
-import org.reflections.Reflections
 
 object SparktaJob extends SLF4JLogging {
 
@@ -66,11 +64,12 @@ object SparktaJob extends SLF4JLogging {
 
   def getPolicyFromZookeeper(policyName: String): AggregationPoliciesModel = {
     Try({
-    val configSparkta = SparktaConfig.initConfig(AppConstant.ConfigAppName)
-    val curatorFramework = CuratorFactoryHolder.getInstance(configSparkta).get
+      val configSparkta = SparktaConfig.initConfig(AppConstant.ConfigAppName)
+      val curatorFramework = CuratorFactoryHolder.getInstance(configSparkta).get
 
-    read[AggregationPoliciesModel](new Predef.String(curatorFramework.getData.forPath(
-      s"${AppConstant.PoliciesBasePath}/${policyName}")))}) match {
+      read[AggregationPoliciesModel](new Predef.String(curatorFramework.getData.forPath(
+        s"${AppConstant.PoliciesBasePath}/${policyName}")))
+    }) match {
       case Success(policy) => policy
       case Failure(e) => log.error("", e); throw e
     }
