+++ serving-api/src/main/scala/com/stratio/sparkta/serving/api/actor/SwaggerActor.scala
@@ -19,10 +19,11 @@ package com.stratio.sparkta.serving.api.actor
 import akka.actor.ActorContext
 import akka.event.slf4j.SLF4JLogging
 import com.gettyimages.spray.swagger.SwaggerHttpService
-import com.stratio.sparkta.driver.models.{ErrorModel, StreamingContextStatusEnum}
 import com.stratio.sparkta.sdk.JsoneyStringSerializer
+import com.stratio.sparkta.serving.api.actor.SwaggerActor._
 import com.stratio.sparkta.serving.api.constants.HttpConstant
 import com.stratio.sparkta.serving.api.service.http._
+import com.stratio.sparkta.serving.core.models.{ErrorModel, StreamingContextStatusEnum}
 import com.wordnik.swagger.model.ApiInfo
 import org.json4s.DefaultFormats
 import org.json4s.ext.EnumNameSerializer
@@ -30,7 +31,6 @@ import org.json4s.native.Serialization._
 import spray.http.StatusCodes
 import spray.routing._
 import spray.util.LoggingContext
-import SwaggerActor._
 
 import scala.reflect.runtime.universe._
