+++ serving-api/src/main/scala/com/stratio/sparkta/serving/api/helpers/SparktaHelper.scala
@@ -24,10 +24,10 @@ import akka.actor.{ActorSystem, Props}
 import akka.event.slf4j.SLF4JLogging
 import akka.io.IO
 import akka.routing.RoundRobinPool
+import com.stratio.sparkta.driver.constants.AkkaConstant
 import com.stratio.sparkta.driver.factory.SparkContextFactory
 import com.stratio.sparkta.driver.service.StreamingContextService
 import com.stratio.sparkta.serving.api.actor._
-import com.stratio.sparkta.serving.api.constants.AkkaConstant
 import com.stratio.sparkta.serving.core._
 import com.typesafe.config.Config
 import spray.can.Http
