+++ serving-api/src/main/scala/com/stratio/sparkta/serving/api/service/http/PolicyHttpService.scala
@@ -19,10 +19,11 @@ package com.stratio.sparkta.serving.api.service.http
 import javax.ws.rs.Path
 
 import akka.pattern.ask
-import com.stratio.sparkta.driver.models._
-import com.stratio.sparkta.serving.api.actor.StreamingActor._
+import com.stratio.sparkta.driver.constants.AkkaConstant
+import com.stratio.sparkta.serving.api.actor.StreamingActor.CreateContext
 import com.stratio.sparkta.serving.api.actor._
-import com.stratio.sparkta.serving.api.constants.{AkkaConstant, HttpConstant}
+import com.stratio.sparkta.serving.core.models._
+import com.stratio.sparkta.serving.api.constants.HttpConstant
 import com.stratio.sparkta.serving.api.helpers.PolicyHelper
 import com.wordnik.swagger.annotations._
 import spray.http.{HttpResponse, StatusCodes}
