+++ serving-api/src/test/scala/com/stratio/sparkta/serving/api/test/actor/StandAloneContextActorSpec.scala
@@ -19,11 +19,11 @@ package com.stratio.sparkta.driver.test.actor
 import akka.actor.{ActorRef, ActorSystem, Props}
 import akka.pattern.ask
 import akka.testkit.{DefaultTimeout, ImplicitSender, TestKit}
-import com.stratio.sparkta.driver.models.StreamingContextStatusEnum._
-import com.stratio.sparkta.driver.models._
 import com.stratio.sparkta.driver.service.StreamingContextService
-import com.stratio.sparkta.serving.api.actor.StreamingActor._
-import com.stratio.sparkta.serving.api.actor._
+import com.stratio.sparkta.serving.api.actor.StreamingActor.{CreateContext, GetAllContextStatus, GetContextStatus}
+import com.stratio.sparkta.serving.api.actor.{StreamingActor, SupervisorContextActor}
+import com.stratio.sparkta.serving.core.models.StreamingContextStatusEnum._
+import com.stratio.sparkta.serving.core.models._
 import com.typesafe.config.ConfigFactory
 import org.apache.spark.streaming.StreamingContext
 import org.junit.runner.RunWith
