+++ serving-core/pom.xml
@@ -35,6 +35,11 @@
 
     <dependencies>
         <dependency>
+            <groupId>com.stratio.sparkta</groupId>
+            <artifactId>sdk</artifactId>
+            <version>0.6.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>org.apache.curator</groupId>
             <artifactId>curator-framework</artifactId>
             <version>${curator.version}</version>
@@ -56,24 +61,13 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.json4s</groupId>
-            <artifactId>json4s-native_${scala.binary.version}</artifactId>
-            <version>3.2.10</version>
+            <groupId>com.github.fge</groupId>
+            <artifactId>json-schema-validator</artifactId>
         </dependency>
         <dependency>
             <groupId>org.json4s</groupId>
-            <artifactId>json4s-ext_${scala.binary.version}</artifactId>
+            <artifactId>json4s-jackson_${scala.binary.version}</artifactId>
             <version>3.2.10</version>
-        </dependency>
-        <dependency>
-            <groupId>org.spark-project.akka</groupId>
-            <artifactId>akka-actor_${scala.binary.version}</artifactId>
-            <version>${akka.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.spark-project.akka</groupId>
-            <artifactId>akka-slf4j_${scala.binary.version}</artifactId>
-            <version>2.3.4-spark</version>
         </dependency>
     </dependencies>
